+++ libs/x-content/src/main/java/org/opensearch/common/xcontent/json/JsonXContent.java
@@ -37,6 +37,7 @@ import com.fasterxml.jackson.core.JsonFactory;
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.core.JsonParser;
 import org.opensearch.common.xcontent.DeprecationHandler;
+import org.opensearch.common.xcontent.MediaType;
 import org.opensearch.common.xcontent.NamedXContentRegistry;
 import org.opensearch.common.xcontent.XContent;
 import org.opensearch.common.xcontent.XContentBuilder;
@@ -77,7 +78,7 @@ public class JsonXContent implements XContent {
     private JsonXContent() {}
 
     @Override
-    public XContentType type() {
+    public MediaType mediaType() {
         return XContentType.JSON;
     }